Update test utils
