+++ packages/playwright/src/tests/04-kunstalte-homepage.spec.ts
@@ -1,12 +1,12 @@
 import { expect, test } from '@playwright/test'
 
-import { getVercelUrl } from '../utils'
+import { getUrl } from '../utils'
 
 test.describe('04. Kunsthalte Homepage', () => {
   // Test the title of the homepage
   test('TC-01: should display translated titles', async ({ page }) => {
     // Go to the homepage
-    await page.goto(getVercelUrl('kunsthalte'))
+    await page.goto(getUrl('kunsthalte'))
 
     //   Check TR title
     await page.click('button:has-text("TR")')
