+++ packages/playwright/src/utils/getUrl.ts
@@ -1,0 +1,11 @@
+import type { Site } from '@fc/types'
+
+import { ports, stagingUrls } from '../config'
+
+export const getUrl = (site: Site) => {
+  if (process.env.CI === 'true') {
+    return stagingUrls[site]
+  }
+
+  return `http://localhost:${ports[site]}`
+}
